third_party/blink/web_tests/external/wpt/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/legend-align-text-align-expected.txt

This is a testharness.js-based test.
[FAIL] legend[align=left]
  assert_equals: expected "start" but got "left"
[FAIL] legend[align=center]
  assert_equals: expected "start" but got "center"
[FAIL] legend[align=right]
  assert_equals: expected "start" but got "right"
[FAIL] legend[align=justify]
  assert_equals: expected "start" but got "justify"
Harness: the test ran to completion.
